+++ lib/data/money_flow.dart
@@ -4,7 +4,7 @@ import "package:flow/data/currencies.dart";
 import "package:flow/data/exchange_rates.dart";
 import "package:flow/data/money.dart";
 import "package:flow/entity/transaction.dart";
-import "package:flow/prefs.dart";
+import "package:flow/prefs/local_preferences.dart";
 
 class MoneyFlow<T> {
   final T? associatedData;