+++ lib/entity/transaction/extensions/default/geo.dart
@@ -1,5 +1,5 @@
 import "package:flow/entity/transaction/extensions/base.dart";
-import "package:flow/utils/utc_datetime_converter.dart";
+import "package:flow/utils/json/utc_datetime_converter.dart";
 import "package:flow/utils/utils.dart";
 import "package:geolocator/geolocator.dart";
 import "package:json_annotation/json_annotation.dart";