+++ lib/prefs/local_preferences.dart
@@ -1,21 +1,21 @@
 import "dart:async";
-import "dart:convert";
-import "dart:developer";
 
 import "package:flow/data/exchange_rates_set.dart";
-import "package:flow/data/prefs/frecency.dart";
 import "package:flow/entity/account.dart";
-import "package:flow/entity/category.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/objectbox.dart";
 import "package:flow/objectbox/objectbox.g.dart";
 import "package:flow/prefs/pending_transactions.dart";
 import "package:flow/prefs/theme.dart";
+import "package:flow/prefs/transitive.dart";
 import "package:flow/theme/color_themes/registry.dart";
 import "package:intl/intl.dart";
 import "package:local_settings/local_settings.dart";
-import "package:moment_dart/moment_dart.dart";
 import "package:shared_preferences/shared_preferences.dart";
+
+export "./pending_transactions.dart";
+export "./theme.dart";
+export "./transitive.dart";
 
 /// This class contains everything that's stored on
 /// device. Such as user preferences and per-device
@@ -57,27 +57,20 @@ class LocalPreferences {
 
   late final LocaleSettingsEntry localeOverride;
 
-  /// Whether the user uses only one currency across accounts
-  late final BoolSettingsEntry transitiveUsesSingleCurrency;
-
-  late final DateTimeSettingsEntry transitiveLastTimeFrecencyUpdated;
-
   late final JsonSettingsEntry<ExchangeRatesSet> exchangeRatesCache;
 
   late final BoolSettingsEntry enableGeo;
 
   late final BoolSettingsEntry autoAttachTransactionGeo;
 
-  late final BoolSettingsEntry requirePendingTransactionConfrimation;
-
   late final BoolSettingsEntry privacyMode;
-  late final BoolSettingsEntry sessionPrivacyMode;
 
   late final BoolSettingsEntry preferFullAmounts;
   late final BoolSettingsEntry useCurrencySymbol;
 
   late final PendingTransactionsLocalPreferences pendingTransactions;
   late final ThemeLocalPreferences theme;
+  late final TransitiveLocalPreferences transitive;
 
   LocalPreferences._internal(this._prefs) {
     SettingsEntry.defaultPrefix = "flow.";
@@ -127,17 +120,6 @@ class LocalPreferences {
       preferences: _prefs,
     );
 
-    transitiveUsesSingleCurrency = BoolSettingsEntry(
-      key: "transitive.usesSingleCurrency",
-      preferences: _prefs,
-      initialValue: true,
-    );
-
-    transitiveLastTimeFrecencyUpdated = DateTimeSettingsEntry(
-      key: "transitive.lastTimeFrecencyUpdated",
-      preferences: _prefs,
-    );
-
     exchangeRatesCache = JsonSettingsEntry<ExchangeRatesSet>(
       initialValue: ExchangeRatesSet({}),
       key: "caches.exchangeRatesCache",
@@ -158,22 +140,12 @@ class LocalPreferences {
       initialValue: false,
     );
 
-    requirePendingTransactionConfrimation = BoolSettingsEntry(
-      key: "requirePendingTransactionConfrimation",
-      preferences: _prefs,
-      initialValue: true,
-    );
-
     privacyMode = BoolSettingsEntry(
       key: "privacyMode",
       preferences: _prefs,
       initialValue: false,
     );
-    sessionPrivacyMode = BoolSettingsEntry(
-      key: "transitive.sessionPrivacyMode",
-      preferences: _prefs,
-      initialValue: false,
-    );
+
     preferFullAmounts = BoolSettingsEntry(
       key: "preferFullAmounts",
       preferences: _prefs,
@@ -191,167 +163,10 @@ class LocalPreferences {
       initialValue: null,
     );
 
-    updateTransitiveProperties();
-
     pendingTransactions =
         PendingTransactionsLocalPreferences.initialize(_prefs);
     theme = ThemeLocalPreferences.initialize(_prefs);
-  }
-
-  Future<void> updateTransitiveProperties() async {
-    try {
-      final accounts = await ObjectBox().box<Account>().getAllAsync();
-
-      final usesSingleCurrency =
-          accounts.map((e) => e.currency).toSet().length == 1;
-
-      await transitiveUsesSingleCurrency.set(usesSingleCurrency);
-    } catch (e) {
-      log("[LocalPreferences] cannot update transitive properties due to: $e");
-    }
-
-    try {
-      unawaited(sessionPrivacyMode.set(privacyMode.get()));
-    } catch (e) {
-      // Silent fail
-    }
-
-    try {
-      if (transitiveLastTimeFrecencyUpdated.get() == null ||
-          !transitiveLastTimeFrecencyUpdated
-              .get()!
-              .isAtSameDayAs(Moment.now())) {
-        unawaited(_reevaluateCategoryFrecency());
-        unawaited(_reevaluateAccountFrecency());
-        unawaited(transitiveLastTimeFrecencyUpdated.set(DateTime.now()));
-      }
-    } catch (e) {
-      // Silent fail
-    }
-  }
-
-  Future<FrecencyData?> setFrecencyData(
-    String type,
-    String uuid,
-    FrecencyData? value,
-  ) async {
-    final String prefixedKey = "transitive.frecency.$type.$uuid";
-
-    if (value == null) {
-      await _prefs.remove(prefixedKey);
-      return null;
-    } else {
-      await _prefs.setString(prefixedKey, jsonEncode(value.toJson()));
-      return value;
-    }
-  }
-
-  Future<FrecencyData?> updateFrecencyData(
-    String type,
-    String uuid,
-  ) async {
-    final FrecencyData current = getFrecencyData(type, uuid) ??
-        FrecencyData(lastUsed: DateTime.now(), useCount: 0, uuid: uuid);
-
-    return await setFrecencyData(type, uuid, current.incremented());
-  }
-
-  FrecencyData? getFrecencyData(String type, String uuid) {
-    final String prefixedKey = "transitive.frecency.$type.$uuid";
-
-    final raw = _prefs.getString(prefixedKey);
-
-    if (raw == null) return null;
-
-    try {
-      return FrecencyData.fromJson(jsonDecode(raw));
-    } catch (e) {
-      return null;
-    }
-  }
-
-  Future<void> _reevaluateCategoryFrecency() async {
-    final List<Category> categories =
-        await ObjectBox().box<Category>().getAllAsync();
-
-    if (categories.isEmpty) {
-      return;
-    }
-
-    for (final category in categories) {
-      try {
-        final Query<Transaction> categoryTransactionsQuery = ObjectBox()
-            .box<Transaction>()
-            .query(Transaction_.categoryUuid.equals(category.uuid).and(
-                Transaction_.transactionDate
-                    .lessThan(DateTime.now().millisecondsSinceEpoch)))
-            .order(Transaction_.transactionDate, flags: Order.descending)
-            .build();
-
-        final int useCount = categoryTransactionsQuery.count();
-        final DateTime lastUsed =
-            categoryTransactionsQuery.findFirst()?.transactionDate ??
-                DateTime.fromMillisecondsSinceEpoch(0);
-
-        categoryTransactionsQuery.close();
-
-        unawaited(
-          setFrecencyData(
-            "category",
-            category.uuid,
-            FrecencyData(
-              uuid: category.uuid,
-              lastUsed: lastUsed,
-              useCount: useCount,
-            ),
-          ),
-        );
-      } catch (e) {
-        log("Failed to build category FrecencyData for $category due to: $e");
-      }
-    }
-  }
-
-  Future<void> _reevaluateAccountFrecency() async {
-    final List<Account> accounts =
-        await ObjectBox().box<Account>().getAllAsync();
-
-    if (accounts.isEmpty) {
-      return;
-    }
-
-    for (final account in accounts) {
-      try {
-        final Query<Transaction> accountTransactionsQuery = ObjectBox()
-            .box<Transaction>()
-            .query(Transaction_.accountUuid.equals(account.uuid).and(
-                Transaction_.transactionDate
-                    .lessThan(DateTime.now().millisecondsSinceEpoch)))
-            .order(Transaction_.transactionDate, flags: Order.descending)
-            .build();
-
-        final int useCount = accountTransactionsQuery.count();
-        final DateTime lastUsed =
-            accountTransactionsQuery.findFirst()?.transactionDate ??
-                DateTime.fromMillisecondsSinceEpoch(0);
-
-        accountTransactionsQuery.close();
-
-        unawaited(
-          setFrecencyData(
-            "account",
-            account.uuid,
-            FrecencyData(
-              uuid: account.uuid,
-              lastUsed: lastUsed,
-              useCount: useCount,
-            ),
-          ),
-        );
-      } catch (e) {
-        log("Failed to build account FrecencyData for $account due to: $e");
-      }
-    }
+    transitive = TransitiveLocalPreferences.initialize(_prefs);
   }
 
   String getPrimaryCurrency() {