+++ lib/routes/home/home_tab.dart
@@ -1,11 +1,11 @@
 import "dart:async";
 
 import "package:flow/data/exchange_rates.dart";
-import "package:flow/data/transactions_filter.dart";
+import "package:flow/data/transaction_filter.dart";
+import "package:flow/data/transactions_filter/time_range.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/objectbox/actions.dart";
-import "package:flow/prefs.dart";
-import "package:flow/prefs/pending_transactions.dart";
+import "package:flow/prefs/local_preferences.dart";
 import "package:flow/services/exchange_rates.dart";
 import "package:flow/utils/utils.dart";
 import "package:flow/widgets/default_transaction_filter_head.dart";
@@ -18,7 +18,6 @@ import "package:flow/widgets/home/home/flow_cards.dart";
 import "package:flow/widgets/home/home/no_transactions.dart";
 import "package:flow/widgets/rates_missing_warning.dart";
 import "package:flow/widgets/transactions_date_header.dart";
-import "package:flow/widgets/utils/time_and_range.dart";
 import "package:flutter/material.dart";
 import "package:moment_dart/moment_dart.dart";
 
@@ -38,7 +37,7 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
   late int _plannedTransactionsNextNDays;
 
   TransactionFilter defaultFilter = TransactionFilter(
-    range: last30Days(),
+    range: TransactionFilterTimeRange.last30Days,
   );
   DateTime dateKey = Moment.startOfToday();
 
@@ -49,14 +48,18 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
         .add(Duration(days: _plannedTransactionsNextNDays))
         .startOfNextDay();
 
-    if (currentFilter.range != null &&
-        currentFilter.range!.contains(Moment.now()) &&
-        !currentFilter.range!.contains(plannedTransactionTo)) {
+    final TimeRange? timeRange = currentFilter.range?.range;
+
+    if (timeRange != null &&
+        timeRange.contains(Moment.now()) &&
+        !timeRange.contains(plannedTransactionTo)) {
       return currentFilter.copyWithOptional(
         range: Optional(
-          CustomTimeRange(
-            currentFilter.range!.from,
-            plannedTransactionTo,
+          TransactionFilterTimeRange.fromTimeRange(
+            CustomTimeRange(
+              timeRange.from,
+              plannedTransactionTo,
+            ),
           ),
         ),
       );
@@ -167,6 +170,9 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
     List<Transaction> transactions,
     ExchangeRates? rates,
   ) {
+    final bool showMissingExchangeRatesWarning = rates == null &&
+        TransitiveLocalPreferences().transitiveUsesSingleCurrency.get();
+
     final Map<TimeRange, List<Transaction>> grouped = transactions
         .where((transaction) =>
             !transaction.transactionDate.isAfter(now) &&
@@ -202,7 +208,7 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
             transactions: transactions,
             rates: rates,
           ),
-          if (rates == null) ...[
+          if (showMissingExchangeRatesWarning) ...[
             const SizedBox(height: 12.0),
             RatesMissingWarning(),
           ],
@@ -251,7 +257,7 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
     setState(() {
       dateKey = Moment.startOfToday();
       defaultFilter = TransactionFilter(
-        range: last30Days(),
+        range: TransactionFilterTimeRange.last30Days,
       );
     });
   }