+++ lib/routes/setup/setup_categories_page.dart
@@ -5,7 +5,7 @@ import "package:flow/entity/category.dart";
 import "package:flow/l10n/extensions.dart";
 import "package:flow/objectbox.dart";
 import "package:flow/objectbox/objectbox.g.dart";
-import "package:flow/prefs.dart";
+import "package:flow/prefs/local_preferences.dart";
 import "package:flow/utils/utils.dart";
 import "package:flow/widgets/add_category_card.dart";
 import "package:flow/widgets/category_card.dart";