+++ lib/widgets/rates_missing_warning.dart
@@ -1,5 +1,5 @@
 import "package:flow/l10n/extensions.dart";
-import "package:flow/prefs.dart";
+import "package:flow/prefs/local_preferences.dart";
 import "package:flow/services/exchange_rates.dart";
 import "package:flow/theme/theme.dart";
 import "package:flow/utils/extensions/toast.dart";