+++ lib/widgets/transaction_list_tile.dart
@@ -1,5 +1,5 @@
 import "package:flow/data/flow_icon.dart";
-import "package:flow/data/transactions_filter.dart";
+import "package:flow/data/transaction_filter.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/entity/transaction/extensions/default/transfer.dart";
 import "package:flow/l10n/extensions.dart";